Show protocol tab if user has no credentials
