+++ src/popup/containers/HomeScreen/index.tsx
@@ -1,8 +1,34 @@
 import { withNavBar } from "@popup/components/common/NavBar";
-import TabBar from "@popup/components/common/TabBar";
+import Tabs from "@popup/components/common/Tabs";
+import Protocol from "@popup/components/Protocol";
 
-function Home() {
-  return <TabBar />;
+import { useAppSelector } from "@redux/stores/application/context";
+import { getProtocolEnabled } from "@redux/stores/application/reducers/app/selectors";
+
+interface HomeScreenProps {
+  credentials: (props: any) => JSX.Element;
 }
 
-export default withNavBar(Home);
+function HomeScreen({ credentials }: HomeScreenProps) {
+  const protocolEnabled = useAppSelector(getProtocolEnabled);
+
+  const tabs = [
+    {
+      id: "credentials-tab",
+      label: "Credentials",
+      props: {},
+      component: credentials,
+    },
+    {
+      id: "protocol-tab",
+      label: "Protocol",
+      props: {},
+      component: Protocol,
+      disabled: !protocolEnabled,
+    },
+  ];
+
+  return <Tabs tabs={tabs} />;
+}
+
+export default withNavBar(HomeScreen);
